+++ src/canvas/wrap.ts
@@ -108,11 +108,14 @@ export function buildWidgetDocument(
   // This bridge precedes widget code and snapshots every authority-bearing
   // primitive. Inline chat keeps its private prompt port; board hosting adopts
   // the view ticket and routes every host API over one request channel.
-  // The click listener is that channel's navigation half: the frame sandbox
-  // grants no popups, so only a trusted user click on a ticketed board widget's
-  // http(s) `target="_blank"` anchor reaches the host. Widening any of those
-  // three gates hands ungranted widgets an outbound channel that the
-  // `connect-src` CSP in board-sandbox.ts otherwise denies.
+  // The activation listeners are that channel's navigation half: the frame
+  // sandbox grants no popups, so only a trusted new-tab activation on a
+  // ticketed board widget's http(s) anchor reaches the host. Widening the
+  // trusted/ticketed/http(s) gates hands ungranted widgets an outbound channel
+  // that the `connect-src` CSP in board-sandbox.ts otherwise denies. The
+  // activation predicate mirrors ui/src/app/native-link-routing.ts so widget
+  // links honor the same primary-click and middle-button contract; listening on
+  // bubble rather than capture keeps a widget's own preventDefault effective.
   const widgetBridge =
     '<script>(()=>{if(!window.parent||window.parent===window||Object.prototype.hasOwnProperty.call(window,"openclaw"))return;' +
     "const parent=window.parent;const post=parent.postMessage.bind(parent);" +
@@ -152,10 +155,14 @@ export function buildWidgetDocument(
     "catch(error){pending.delete(id);reject(error);}};" +
     'if(ticket)send();else if(hostInitExpired)reject(new ErrorCtor("widget host capabilities unavailable"));' +
     "else push.call(waiting,{send,reject});});" +
-    'listen("click",event=>{if(event.isTrusted!==true||!ticket||event.defaultPrevented||event.button!==0)return;' +
+    "const activate=event=>{if(event.isTrusted!==true||!ticket||event.defaultPrevented||event.shiftKey||event.altKey)return;" +
+    'const auxiliary=event.type==="auxclick"&&event.button===1;' +
+    'if(!auxiliary&&!(event.type==="click"&&event.button===0))return;' +
     'for(let index=0,entries=path.call(event);index<entries.length;index++){const anchor=entries[index];if(anchor?.tagName!=="A")continue;' +
-    'const url=anchor.href;if(!url)continue;if(anchor.target!=="_blank"||!test.call(/^https?:\\/\\//i,url))return;' +
-    'prevent.call(event);then.call(request("host.open",{url}),()=>{},()=>{});return;}},{capture:true});' +
+    'const url=anchor.href;if(!url)continue;if(!auxiliary&&anchor.target!=="_blank")return;' +
+    "if(!test.call(/^https?:\\/\\//i,url))return;" +
+    'prevent.call(event);then.call(request("host.open",{url}),()=>{},()=>{});return;}};' +
+    'listen("click",activate);listen("auxclick",activate);' +
     "const sendPrompt=text=>{if(!act||act()!==true)return P.resolve(false);const value=stringify(text);" +
     'if(ticket)return request("prompt.send",{text:value});return new P((resolve,reject)=>{' +
     'const send=()=>{const result=request("prompt.send",{text:value});then.call(result,resolve,reject);};' +
